+++ src/main/java/com/dreamsportslabs/guardian/constant/Constants.java
@@ -10,7 +10,7 @@ public final class Constants {
 
   public static final String TOKEN = "token";
   public static final String CODE = "code";
-  public static final String JWKS_KEYS = "jwks_keys";
+  public static final String JWKS_KEYS = "keys";
 
   public static final String FACEBOOK = "facebook";
   public static final String GOOGLE = "google";
