+++ src/main/java/com/dreamsportslabs/guardian/service/JwksService.java
@@ -2,7 +2,9 @@ package com.dreamsportslabs.guardian.service;
 
 import com.dreamsportslabs.guardian.cache.TenantCache;
 import com.dreamsportslabs.guardian.config.tenant.TokenConfig;
-import com.dreamsportslabs.guardian.utils.KeyUtils;
+import io.fusionauth.jwks.domain.JSONWebKey;
+import io.fusionauth.jwt.domain.Algorithm;
+import io.fusionauth.jwt.domain.KeyType;
 import io.reactivex.rxjava3.core.Single;
 import io.vertx.core.json.JsonObject;
 import java.util.List;
@@ -23,8 +25,16 @@ public class JwksService {
               TokenConfig tokenConfig = tenantConfig.getTokenConfig();
               String alg = tokenConfig.getAlgorithm();
               return tokenConfig.getRsaKeys().stream()
-                  .map(key -> KeyUtils.getKeysInJwksFormat(key.getPublicKey(), key.getKid(), alg))
+                  .map(key -> getKeysInJwksFormat(key.getPublicKey(), key.getKid(), alg))
                   .collect(Collectors.toList());
             });
+  }
+
+  public static JsonObject getKeysInJwksFormat(String publicKey, String kid, String alg) {
+    JSONWebKey jwk = JSONWebKey.build(publicKey);
+    jwk.kid = kid;
+    jwk.kty = KeyType.RSA;
+    jwk.alg = Algorithm.fromName(alg);
+    return new JsonObject(jwk.toJSON());
   }
 }
